made distortion array D resizable
git-svn-id: http://svn.code.sf.net/p/bosch-ros-pkg/code/trunk/stacks/bosch_drivers/usb_cam@552 2f5dae86-743c-430d-921f-7dbcfe3428eb

+++ src/usb_cam_node/usb_cam_node.cpp
@@ -96,7 +96,7 @@ public:
 
       if ((double_list.getType() == XmlRpc::XmlRpcValue::TypeArray)) {
         info_.D.resize(double_list.size());
-        for (size_t i=0; i<double_list.size(); i++) {
+        for (int i=0; i<double_list.size(); i++) {
           ROS_ASSERT(double_list[0].getType() == XmlRpc::XmlRpcValue::TypeDouble);
           info_.D[i] = double_list[i];
         }
